Rename cute_sketch to contest_painting_effects, and battle_1 to decompress
