+++ asm/contest_painting_effects.s
