+++ asm/decompress.s
