+++ data/contest_painting_effects.s
