+++ include/battle.h
@@ -708,7 +708,7 @@ void MarkBufferBankForExecution(u8 bank);
 
 extern u8 gBattleTextBuff1[];
 
-// asm/battle_1.o
+// src/battle_bg.o
 void sub_800D6D4();
 void ApplyPlayerChosenFrameToBattleMenu();
 void DrawMainBattleBackground(void);
