+++ include/contest_painting_effects.h
